PS01/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
spacy==3.8.12
mem0ai==0.0.15
ollama==0.6.1
docker==7.0.0
python-dotenv==1.0.0
redis[asyncio]>=5.0.0
aiokafka>=0.10.0
pytest==7.4.3
pytest-asyncio==0.21.1
pydantic==2.9.0
